Guard against missing data
If there is any chance post.user could be null (e.g., user deleted), we
should capture the prop and handle it in the script.  By doing this, we
can prevent runtime errors or blank spots if data is incomplete.

Vue does not allow mixing type parameters (generics) and runtime
validation objects in a single defineProps call.  Therefore we use the
runtime as it outweighs the type parameter.