Build auto hide for flash messages
We don't need to change the FlashMessages interface as it's shared and
strictly defined.  Each type uses | null and all fields are required.
Populate all its members with null by default for compatibility without
changing the interface

Use a local ref to prevent prop mutation.  To handle rapid submits, if a
new message arrives in fewer than 5s, the old timeout is canceled.

Show messages right away via immediate: true and the existing timeout is
cleared on unmount to make it memory-safe.